renamed run command to daemon
